+++ core/src/index.ts
@@ -19,7 +19,7 @@ export * from "./perspectives/PerspectiveProxy";
 export * from "./perspectives/PerspectiveDiff";
 export * from "./perspectives/LinkQuery";
 export * from "./SmartLiteral";
-export * from "./subject/SDNADecorators";
+export * from "./subject/decorators";
 export * from "./subject/Subject";
 export * from "./neighbourhood/Neighbourhood";
 export * from "./neighbourhood/NeighbourhoodProxy";
